Refactor Middleware to a single set of functions for all routes  (#3534)
The multiple sets of Middleware functions was getting confusing and
difficult to reason about. Now there is a single set of Middleware
functions that run for every route before the handler is dispatched.

A route can provide an OptionsOverride function to change the
middleware.Options passed to the Middleware for a route. This
allows a route to tweak how the Middleware runs for a route (e.g
response logging).

The set of Middleware is completely configurable on the
HandlerOptions. If not set, a set of Default middleware is used.

A CustomHandler can also provide an OptionOverride function, in addition
to the OptionOverride already set on the previous handler. This allows
further configuring the options for the CustomHandler.