Explicitly create static or dynamic pooled worker pools
This refactor requires code that uses a worker pool to explicitly choose
static or dynamic. It has been proven to be quite buggy/error prone to
allow a configurable worker pool type, because they are quite different.

1. a dynamic worker pool should never block, so it should not have
   methods like GoWitTimeout since it can never timeout.
2. test code was using a static worker pool, when prod code would be
   using a dynamic worker pool.
3. it can be quite dangerous to allow switching an unbounded pool to a
   bounded pool in prod.

Additionally, buffer channels were removed from the static worker pool.
Allowing a buffer can potentially lead to incorrect code when there is a
slot available, but the worker does not process it until after the
timeout. The client would have expected the enqueue request to timeout
instead. Buffered channels were removed from the dynamic queue earlier
this year for similar buggy reasons.

Removing the buffer made reasoning about the size much easier. Now there
is a single option, NumShards, to configure the number of worker
channels. This is the same for both static and dynamic. The only
difference with the dynamic is additionally goroutines are spun up
temporarily to handle bursts of traffic.