Reuse TagTransformer to avoid heavy memory allocations
Previously a new function to transform tags would be allocated per
datapoint written to the database. This can result in a large number of
allocations for a high throughput system.

Now a TagTransformer is allocated per write operation (which is already
pooled) and can be reset when the write operation is given back to the
pool.