[query] Support rewriting range duration in queries  (#3579)
This commit gives users the ability to configure whether they'd like to rewrite range durations in queries that would be routed to namespaces that have resolutions greater than the range. When disabled, behavior stays the same as it currently is today. Namely, if the range duration is less than the resolution on the namespace, the query will return no data. When enabled and this condition is detected, the range within the query will automatically be updated to a multiple of the largest namespace resolution for the query. That multiple is configurable. It defaults to 0, which turns the rewriting behavior off.

Co-authored-by: Rob Skillington <rob@chronosphere.io>