[query] Sort by ids in ID dedupe map list method (#3635)
The change fixes a bug in how we dedupe series on the graphite query path. In particular the `list()` function on `idDedupMap` does not return deterministic ordering of entries. The change sorts the series by id before returning, so multiple calls to the `list()` method always return things in the same order.