Move addToReset migration to the aggregationKey.Pipeline (#3668)
This will allow removing the migration in the future. This is a noop
change.

At the moment the aggregationKey still contains Adds instead of Resets,
which makes it impossible to seamlessly transition the rollup rules to
explicitly use Reset instead of Add, since the AggregationKey will change.
If the AggregationKey changes it will result in split aggregations
during the deploy.

Once this has been deployed to the aggregator, the rollup rules in the
coordinator can be changed to use Reset explicilty and the
AggregationKey won't change in the Aggregator. The boolean can then be
dropped once all rules are using Reset instead of Add.

When this change rolls out to the aggregator it will change the
AggregationKey for incoming metrics. However, it will only affect the
follower and once the follower changes to the leader, the new
AggregationKey is fine to use.