[aggregator] Add interrupt channel for Aggregator shutdown.
Instead of watching for interrupts directly within the
server.Run method. Allow for an interruptCh to be passed
as a RunOption. This will allow the aggregator to be started
and shutdown programatically which will be useful as
we improve upon our integration tests. This pattern is already
used for both the coordinator and db nodes.