[dbnode] Expose single method for starting dbnode and optional coordinator (#3678)
Before this change, the main() method housed logic that determined if
a coordinator needed to be started alongside the dbnode. To ensure
integration tests don't duplicate this logic, we add a wrapper method,
server.RunComponents, that both main() and integration tests can call.

Additionally, allow clients to provide an interrupt channel so that
components can be stopped programmatically within a test.