[dbnode][aggregator][coordinator] Add shutdownCh for server shutdown notification
This commit allows clients to provide a channel to receive
a notification that a component server has shutdown. The
primary use case for this will be for integration tests
that start components in-process. This will allow tests
to know precisely when a server has stopped instead of
relying on polling API endpoints or waiting for a
timeout.