Correctly handle different error types in agg tcp server
The previous code was incorrectly type switching on type aliases, which
does not work as "expected". All errors would result in untimed errors,
since it was the first type alias in the switch statement.

Instead of introducing new error types, just used the existing message
type to switch on the log statements.