Do not set Aggregation for AutoMappingRule
Previously the AutoMappingRule used LAST, which only works for Gauges.
This is fine for Prometheus metrics, which are all Guages. However, LAST
is not a valid aggregation for Counters, which is used by StatsD.

Instead of explicitly setting an Aggregation, the default for the metric
type is used by the Aggregator when it's not set. So LAST will continue
to be used for Gauges and SUM will be used for Counters.