[aggregator][dbnode][query] Improve resource cleanup flow on shutdown (#3696)
* Ensure shutdown notication sent after resource cleanup

This commit moves the shutdown notification to a deferred
function, and ensures that the function is the last deferred
function called by registering it first. This ensures that
all cleanup behavior happening in deferred functions are
complete before we send a shutdown notification.

* Ensure dbnode debug server shutdown on server shutdown

This commit calls server.Shutdown for the debug server
once the main dbnode server is shutdown. This change makes
it easier to restart a DB node process in-memory without
causing errors.