[query] Add M3-Metadata-Count header for tracking fetched metadata
This commit adds an M3-Metadata-Count header that tracks the
amount of metadata (e.g. label names, label values) fetched
from the database before results are potentially truncated
and sent to the browser. This header is analogous to M3-Series-Count
which tracks the same thing but for timeseries instead of metadata.