[query] Add support for classifying queries by result size and time range.
This commit adds the ability to generate coordinator request
metrics tagged with a results and duration bucket. Clients can
configure the number of buckets and the values of those buckets.
Once enabled, each request to supported routes (prom query and label
endpoints) will be analyzed for the number of results returned
and the time window of the request and grouped into the closest
bucket specified in the config. Those buckets will then be
added as a tag on the coordinator request metrics.