Switch from AddUntimed to AddTimed for ResendEnabled
This allows for a graceful transition from untimed -> timed metrics.

Previously when the coordinator would switch from untimed -> timed, it
would result in a new aggregation for that metric on the leader. That
incomplete aggregation would get written by the leader, resulting in
incorrect aggregated data. An example is a prom counter that is not
increasing. The new aggregation starts at 0 and the prom counter might
be at value 1000. Even though the counter hasn't increased, the
aggregator will output an increase of 1000 for the first aggregation,
which is incorrect.

Now the transition happens in the aggregator. When the follower is
deployed with the feature flag enabled, it will migrate untimed ->
timed. Once the follower is ready to take over as the leader, it will
have complete timed aggregations to start flushing.

This is a very similar problem for deploys. The first metrics written to
the follower after a deploy are incomplete, which is why we need to wait
for the bufferDurationForPastTimed metrics for the follower to takeover.