Reuse Metadata slices when Unmarshaling the protos. (#3732)
* Move generated Unmarshal functions to manual file

This is a no-op change in preparation for a future commit that will
change the Unmarshal functions to reuse slices.

* Use value types for pipeline ops

This will allow reusing the memory for the pipeline ops when the
metadata proto is reused for Unmarshaling.

* Reuse Metadata slices when Unmarshaling the protos.

This provides a significant allocation saving. Our tests showed that 50%
of the alloc space was Unmarshaling StagedMetadata. By reusing the
slices of the previous proto, this has completely removed the Unmarshal
fn from the alloc flamegraph when pprof is used.