[dbnode] Allow interrupt when waiting for initial placement.
DB nodes will block indefinitely when waiting for the
placement/topology with no way to interrupt the process.
This is generally fine, but ends up being problematic
during integration tests when we may want to quickly spin
up and spin down a node for various reasons. As such, thread
through the interrupt channel so that we also check for
interrupts while waiting for the placement. If an interrupt
is received, then gracefully shutdown the server.