[query] Allow interrupt when watching during downsampler creation.
Coordinators will block for a default of 10 seconds to wait for
namespace information when creating a downsampler.
This is generally fine, but ends up being problematic
during integration tests when we may want to quickly spin
up and spin down a coordinator for various reasons.
As such, thread through the interrupt channel so that we
also check for interrupts while waiting for namespaces.
If an interrupt is received, then gracefully shutdown the server.