[dbnode,query] Ensure only single listener for interrupt channel
Passing the interrupt channel to multiple goroutines could
cause a race where the main thread ends up missing the interrupt
that triggers a server shutdown. This commit ensures that only
a single goroutine is listening for the interrupt at a given
time and all other interested parties can check the interrupted
channel. The interrupted channel will be closed as soon as
an interrupt is received. Since closed channels return immediately,
this allows any interested goroutine to know if it should terminate
by simply checking the interrupted channel.