[tests] Add support for inprocess.NewCluster (#3797)
* [dbnode] Allow global index claims manager check to be disabled for tests

This check does not allow us to start multiple dbnodes in-process due
to it's global nature. As such, allow it to be disabled and indicate
that this is only safe for tests.

* [tests] Add support for inprocess.NewCluster

inprocess.NewCluster allows clients to quickly spin up
a new cluster based on the desired config. Additionally,
can tweak the cluster topology from a set of ClusterOptions.

As a part of this refactor how we update configuration for
in-process components.