[tests] Add support for resources.ExternalResources (#3811)
This commit adds an interface for adding external (i.e. not M3)
resources that can be used in integration tests. An
implementation of the interface that allows users to spin up
a docker-backed Prometheus is added as well. This also functions
as a working example for how the interface can be implemented to
support other external resources.