[rules] Correctly match rollup tags for excludeBy operations (#3865)
Previously, excludeBy rule evaluation would sometimes skip rules
when in the input metric was missing an exact match for an exclude
rule. This fixes iteration to properly merge the input rule tags
against the exclude rules so that the output only contains tags
that have no corresponding match in the exclude set.