[dbnode] Remove additional indexed for block start checks
This commit removes logic that iterates over the query range
in block size chunks checking to see if the document is indexed
for that block start. This logic is unnecessary as the blocks
that we're processing are already guaranteed to be within the
query range as the blocks iterator is created with the query
range as its bounds. Additionally, since query results are scoped
to a specific block, the results are guaranteed to be
indexed for that block start as well. Therefore, the only
check we need to do is see if the document still exists in
the reverse index for the block start.