Continue using Untimed metrics for resendEnabled pipelines (#3870)
This allows seamlessly transitioning to resendEnabled for a pipeline
without having to do a deploy of the aggregators.

Since timed vs untimed is part of the MetricMap.Key, a new aggregation
is created when switching from Untimed -> Timed. This results in
incorrect data when switching a rule to resendEnabled. Previously we
relied on the failover process to only switch over when a follower was
initially coming up. This worked, but it required a deploy which we want
to avoid.

This puts the entire feature rollout on the Pipeline itself.