Fix race when checking for dirty aggregations (#3886)
This fixes the following race:

1. Writer creates a new aggregation and adds it to the drity set.
2. Flusher processes dirty set.
3. Writer updates the aggregation value.

This race was introduced with the recent refactor of resending to use to
an explicit dirty set. However, this race existed historically before
resending introduced the notion of a dirty bit.

The fix is to ensure the individual aggregation lock is held when
read/writing the dirty bit.

It was starting to get really confusing what state truly needs to be guarded by the
aggregation lock. This also refactors the code to introduce an
aggFlushState of flusher local state that doesn't need any
synchronization. Now all fields on the lockedAggregation must be
accessed with the lock.