Keep pointers to next/prev aggregations
This should help when having to iterate through the map to expire. There
are worst case scenarios when the map is sparse that burn a lot of CPU
doing a lot of empty lookups. This worst case happen when metrics become
dormant or infrequent, and you have to scan the entire "range" of the
map for no reason.