Only keep the bucket version in the forward writer
This minimizes how much data the forward writer needs to keep around
after a flush. Only the version needs to be kept around for the life
time of the the timed aggregation to support resending data.

Additionally, simplified the memory reuse to get rid of the
cachedValues. This is unnecessary and we can just reuse the slices
correctly when adding new aggregations/values.