Fix range rewrite middleware failing to rewrite queries w/ offset
This adds a pointer to the PrometheusEngine to the range rewrite middleware.
We use this engine instance to run the query against a fake Queryable instance
in order to have Prometheus calculate the actual start/end time of the query for us. This is an optional parameter to the middleware and we fall back to using the start/end time on failure.

- Add related query options to QueryStorageMetadataAttributes