[m3msg] Add receive and handle latency to consumers (#3920)
Prior to this commit there was only total end-to-end ack latency on the
producer side. This made it difficult to determine what part of the
end-to-end processing is taking the most time. This commit now breaks
down the end-to-end latency into 2 new timings:
1. receive-latency: how long it took for the consumer to read the
   message after the producer sent it. this includes all network delays
and how long it sat in the consumer receive buffer.
2. handle-latency: how long it took for the consumer to process the
   message after reading it.

There is still one timing missing, how long it takes the producer to
receive the ack message. At the moment it's probably not worth
adding this since it can be derived from the total.