[agg] Fix forward lag calculation (#3924)
When calculating the forwarding lag, we want to add the latenesAllowed
and jitter to the timestamp so it's not counted in the processing lag
number. By subtracting it we were actually adding additional lag that
was incorrect.

forward lag = now - (timestamp + intentional delay)

intentional delay = allowedLateness + jitter

jitter = how long we intentionally delay a flush task so tasks don't run
at the same time.

allowedLateness = how long we intentionally delay a flush task so
allow for late metrics to arrive.