[agg] Fix race when migrating to resendEnabled and the agg is closed  (#3930)
* refactor to move resendEnabled -> lockedAggregation

* [agg] Fix race when migrating to resendEnabled and the agg is closed

This fixes a race when a rollup rule is migrated to resendEnabled, which
triggers using the client timestamp instead of the server timestamp.
However, the client timestamp might be targeting an old aggregation that
is already closed. When this happens, just continue using the
serverTimestamp and target an open aggregation.

This also refactors the existing logic of rolling forward to the next
aggregation when an aggregation is closed. Now that we have multiple
cases, it's easier to handle all of this logic at the higher level.