[msg] Do not Close singleton MessageProcessors when closing connections
This fixes a regression that was introduced in https://github.com/m3db/m3/pull/3918

The shared singleton MessageProcessor was being closed when a Connection
was being closed.

Now a Pool interface is introduced. A MessageProcessor is fetched when a
Connection is created and returned when a Connection is closed. It is up
to the pool impl to decide if it should Close the MessageProcessor on
each return or only when the pool is closed.