[agg] Calculate forwarding lag using expected processing time
Previously the timestamp of the aggregation was used to calculate the
forwaridng lag. This could be misleading for late metrics updating
resendEnabled aggregations, since they are in the past.

Now the lastUpdatedAt time is used. The lastUpdatedAt is used to
calculate the expectedProcessingTime as the next resolution window after
the lastUpdatedAt.