Remove m3collector (#3962)
The m3collector is a dead project and never graduated out of the alpha stage. It doesn't seem to have been updated in over 4 years, but people still have to make refactoring changes to it when they are upgrading some other part of m3. We should remove this dead code to increase development velocity.