[serialize] Do not decode an empty slice
Currently clients call Decoder.Reset(nil) when they are adding the
Decoder back a pool. This always fails and sets the Err. This doesn't
really matter since when the Decoder is fetched back from the pool,
Reset is called again which clears the Err.

A future commit is going to use a TagDecoder without a Pool, so this
needs to be fixed.