[serialize] Add a fast unchecked tag decoder
This adds an unchecked tag decoder/iterator which removes all safety
checks (i.e ref checking) to improve the performance. It is safe to use
this when the underlying bytes are not shared and are safe to reuse.

A good example of this is the downsampler, that is single threaded and
has to match a metric id against many rules. This significantly
increases the performance of the downsampler.