[agg] Do not use uber/multierr
the various m3 errors don't work well with uber/multierr.

For example if you do m3errors.Is(uberMultierr, target) and uberMultierr
contains a m3.InnerError, it won't find it.

uber/multierr is probably better but we'll need to first retrofit a lot
of the m3.InnerError types to support Is(err).