[coordinator] Allow MatchResult to be re-used during rule matching
Creating a new MatchResult each time we want to generate rule
matches for a write has shown to be problematic via profiling.
More specifically, creating MatchResults this way requires us
to allocate O(rule) pipeline slices that end up escaping to the
heap. This ends up causing more frequent GCs which drive up
CPU utilization.

To rememdy this, a MatchResult is now provided to rule matching
code which allows us to make use of the same MatchResult. This
change required us to do any merging of MatchResult data on the
final MatchResult object instead of intermediate structures (
which were the source of the allocations), so a big part of this
change is reworking how we merge result data into the final
MatchResult.