[tests] Ensure aggregation integration test uses remote aggregators (#3996)
* [tests] Separate component creation from start up

This commit allows integration test users to configure
whether they want to start the component or not when component
is created. This is useful as sometimes the entire cluster cannot
be created until something is configured. For example, when
using remote aggregators, the coordinator and aggregators will
fail to come up unless placement data is in etcd. This requires
us to be able to first create the components so we generate ports
which we can then use to create placements and finally start
the components.

* [tests] Ensure aggregation integration test uses remote aggregators