[commitlog] Add setter functions for the onFlush callback (#4045)
* [commitlog] Add setter functions for the onFlush callback

This commit adds a setOnFlush method to both the commitLogWriter and chunkWriter intefaces.

This function is necessary to get around a subtle bug in the WriteWait commitlog strategy
causing pendingFlushFns functions (and therefore commitLog.Write returning) to lag.

The issue is that after the first call to RotateLogs, when onFlush is called, it ends up reading
from the wrong list of pendingFlushFns. This is hard to detect as this occurs after the data has
already been flushed to the correct file on disk and only affects reporting completion.

It is not caught by any tests because the next call to RotateLogs (or Close) will resolve the problem.
Additionally, we mostly test the default WriteBehind strategy rather than the WriteWait strategy.

* Add comment to setOnFlush

* linter fixes

* Empty commit to trigger build