Allow the commitlog package to build on windows (#4065)
Currently, M3 does not run on Windows. The src/dbnode/persist/fs/commitlog package specifically
does not build because of the the internal packages it depends on transitively (src/x/mmap)
uses syscalls that are not available on Windows.

This commit creates a dummy implementation of that package only for Windows allows
the commitlog to be used as a package in binaries built for Windows.