Upgrade dependencies: grpc => 1.44.0, etcd => 3.6.0-alpha.0, otel => 1.4.1
This diff upgrades a number of interrelated dependencies. The goal: upgrade GRPC to 1.44.0.
The fallout: etcd needs to be upgraded to 3.6.0-alpha.0

Gory details:

- We import etcd/server/v3 via etcd/embed and etcd/testing/v3/frameworks/integration.
- etcd/server/v3 in 3.5.2 depends on pre 1.0 opentelemetry. Bleeding edge etcd depends on 1.0 opentelemetry
- M3 depends on 1.0 opentelemetry — this conflicts with etcd 3.5.2, but not bleeding edge etcd

Pinning etcd to 3.6.0-alpha makes things build properly… but of course it’s  a risky move,  since we’re no longer on a release version.
Validations we can do on Uber side (lmk  if more would be helpful):

Bring in the etcd upgrade on a test build (I’m thinking indexer + collector as the ultimate test); bake for a bit.

**Associated code changes**

- All etcd packages are now <package>/v3 instead of packagev3 (e.g. clientv3 => client/v3)
- The GRPC client interfaces changed slightly for the query/remote client. Behavior should be the same (static endpoints with round robin); validated by tests.