[commitlog] Add the ability to configure pools before allocation (#4072)
* Add the ability to configure pools before allocation

There are several places where we allocate large pools of memory
in constructors and provide setters to override those pools.
In general, this is fine as it allows us to provide larger or smaller
pools. However, in super memory constrained settings, we prefer to
avoid the initial alloc and GC all together and provide
the properly sized pools to the constructor.

This PR provides the ability to do this for certain properties
of fs.Options and commitlog.Options.