Don't require m3 repo to be under $GOPATH for code generation (except protobuf)
Since m3 is now a go.mod project, it's no longer necessary to be checked out under GOPATH.

This PR makes our code generation reflect that fact. I've switched to use of relative paths in generate.go, which as I understand
it is the preferred way of doing things generally.

Note: I haven't yet fixed the protobuf generation--there are a couple more kinks to work out there. This PR handles thrift and mocks.