etcd_docker 7: move etcd dependency down to 3.5.4 (#4151)
This changes our etcd version *down* to 3.5.4 from 3.6.0-alpha.0. In https://github.com/m3db/m3/pull/4087, we made the choice to go to
an alpha version of etcd in order to get around the thorny dependency issues described in that ticket. This was partially on the assumption that
etcd would release a full 3.6.0 version in the months following. As of 9/2/2022, that hasn't happened--no progress seems to have been made on etcd
3.6.0.

As described in https://github.com/m3db/m3/issues/4144, we've since addressed the dependency issues via other means, removing our dependency on the
`embed` and `integration` packages. This allows us to move back to an etcd stable version, 3.5.4. Since stable versions are generally preferable,
I'm taking the opportunity to do so.

commit-id:720abd40