[aggregator] ensure quantiles are sorted in stream
Quantile slice that was passed into the stream's ResetSetData was based on the
enum ordering, which places p25 and p75 after all other enums (because p25 and
p75 were added later on in m3 lifetime). But the `Quantile` function expects the
quantile slice to be sorted when fetching a quantile value for the input
quantile.