Fix number of files calculation in test (#4272)
This test was consistently failing for me when run on linux, most likely not due to any issue with the code, but rather to some small change with the number of files opened by the process (kernel update? Who knows).

Better I think to just use the reference implementation to establish a baseline, and then check that we capture the delta. Perfect? No. Good enough? Probably.